Update CI/CD to anticipate, not install or test uncertainties
For now disable testing of `uncertainties` specifically, as that also requires an uncertainties-aware `pint`.  When that becomes available we can update the CI/CD scripts to enable optional uncertainty testing with versions of pint and pint-pandas that can tolerate it.

From local testing with uncertainties, remember: if you don't put a fixture XYZ in the parameter list, then testing `if XYZ` is really testing whether the fixture's function is not None (it's always not None).  When XYZ is in the parameter list, testing `if XYZ` tells you whether the value of XYZ fixture is non-False.

Also increase version compatibility by using `np.array([f(x) for x in arr])` instead of `arr.map(f)` when `arr` is a numpy array.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>